+++ hybris/platforms/common/nativewindowbase.h
@@ -87,7 +87,7 @@ protected:
 	//perform interfaces
 	virtual int setBuffersFormat(int format) = 0;
 	virtual int setBuffersDimensions(int width, int height) = 0;
-	virtual int setUsage(int usage) = 0;
+	virtual int setUsage(uint64_t usage) = 0;
 	virtual int setBufferCount(int cnt) = 0;
 private:
 	static int _setSwapInterval(struct ANativeWindow* window, int interval);
